% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births_per_capita.R
\name{births_per_capita}
\alias{births_per_capita}
\title{births_per_capita}
\usage{
births_per_capita(
  births = NA,
  pop = NA,
  save_to_file = TRUE,
  save_folder = here::here("output_data")
)
}
\arguments{
\item{births}{data}

\item{pop}{data}

\item{save_to_file}{TRUE/FALSE}

\item{save_folder}{string}
}
\value{

}
\description{
includes denominator for
}
\details{
- whole population
 - women only
 - women of childbearing age 15 to 45 years old
}
