% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_births.R
\name{clean_births}
\alias{clean_births}
\title{clean_births}
\usage{
clean_births(
  dir_path = system.file("extdata", "Leeds1", "Births", package = "cleanETHPOP"),
  m_to_100_f = 105,
  rtn = TRUE,
  save_to_file = TRUE,
  save_name = "clean_births",
  save_folder = here::here("output_data"),
  save_format = "csv"
)
}
\arguments{
\item{dir_path}{raw data location}

\item{m_to_100_f}{relative population sizes at birth from ONS}

\item{rtn}{TRUE/FALSE}

\item{save_to_file}{TRUE/FALSE}

\item{save_name}{string}

\item{save_folder}{output data location}

\item{save_format}{csv or RData}
}
\value{

}
\description{
clean_births
}
\examples{
  clean_births()

}
