% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_deaths.R
\name{clean_deaths}
\alias{clean_deaths}
\title{clean_deaths}
\usage{
clean_deaths(
  dir_path = system.file("extdata", "Leeds1", "Deaths", package = "cleanETHPOP"),
  rtn = TRUE,
  save_to_file = TRUE,
  save_name = "clean_deaths",
  save_path = here::here("output_data"),
  save_format = "csv"
)
}
\arguments{
\item{dir_path}{location of ETHPOP data}

\item{rtn}{TRUE/FALSE}

\item{save_to_file}{TRUE/FALSE}

\item{save_name}{output data file}

\item{save_path}{output data location}

\item{save_format}{e.g. csv or RData}
}
\value{

}
\description{
clean_deaths
}
\examples{
  clean_deaths()

}
