% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_inmigrants.R
\name{clean_inmigrants}
\alias{clean_inmigrants}
\title{clean_inmigrants}
\usage{
clean_inmigrants(
  dir_path,
  rtn = TRUE,
  age_grp = FALSE,
  save_to_file = TRUE,
  save_name = "clean_inmigrants",
  save_folder = here::here("output_data"),
  save_format = "csv"
)
}
\arguments{
\item{dir_path}{string file address. this is either Leeds1 or Leeds2.}

\item{age_grp}{}
}
\value{

}
\description{
clean_inmigrants
}
\examples{
  clean_inmigrants()

}
