% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_outmigrants.R
\name{clean_outmigrants}
\alias{clean_outmigrants}
\title{clean_outmigrants}
\usage{
clean_outmigrants(
  dir_path = system.file("extdata", "Leeds1", "OutMigrants", package = "cleanETHPOP"),
  rtn = TRUE,
  age_grp = FALSE,
  save_to_file = TRUE,
  save_name = "clean_outmigrants",
  save_folder = here::here("output_data"),
  save_format = "csv"
)
}
\arguments{
\item{age_grp}{}
}
\value{

}
\description{
clean_outmigrants
}
\examples{
  clean_outmigrants()

}
