% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_pop.R
\name{clean_pop}
\alias{clean_pop}
\title{clean_pop}
\usage{
clean_pop(
  dir_path = system.file("extdata", "Leeds1", "Population", package = "cleanETHPOP"),
  rtn = TRUE,
  age_grp = FALSE,
  save_to_file = TRUE,
  save_name = "clean_pop",
  save_folder = here::here("output_data"),
  save_format = "csv"
)
}
\arguments{
\item{save_format}{}
}
\value{

}
\description{
clean_pop
}
\examples{
  clean_pop()

}
