% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deaths_per_capita.R
\name{deaths_per_capita}
\alias{deaths_per_capita}
\title{deaths_per_capita}
\usage{
deaths_per_capita(
  deaths = NA,
  pop = NA,
  save_to_file = TRUE,
  save_folder = here::here("output_data")
)
}
\arguments{
\item{save_to_file}{}
}
\value{

}
\description{
doesnt join on year 2061
}
