% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outmigrants_per_capita.R
\name{outmigrants_per_capita}
\alias{outmigrants_per_capita}
\title{outmigrants_per_capita}
\usage{
outmigrants_per_capita(
  outmigrants = NA,
  pop = NA,
  save_to_file = TRUE,
  save_folder = here::here("output_data")
)
}
\arguments{
\item{save_to_file}{}
}
\value{

}
\description{
outmigrants_per_capita
}
